requirements.txt
paramiko
flask
flask-cors
requests
python-dotenv
pyserial
escpos
psutil
pyusb